+++ lib/kvstore.js
@@ -84,6 +84,7 @@ var AVAILABLE_BACKENDS = DEFAULT_OPTIONS.available_backends.reduce(
 
 
 module.exports = {
+  ERROR_CAS_MISMATCH: 'cas mismatch',
 
   connect: function(options, cb) {
     options = Hapi.utils.applyToDefaults(DEFAULT_OPTIONS, options || {});
