+++ lib/kvstore/couchbase.js
@@ -10,6 +10,7 @@
 var couchbase = require('couchbase');
 var config = require('../config');
 var Hapi = require('hapi');
+var kvstore = require('../kvstore');
 
 
 module.exports = {
@@ -54,7 +55,7 @@ module.exports = {
         // logic to be aplied after either.
         function handler(err) {
           if (err && err.code === couchbase.errors.keyAlreadyExists) {
-            return cb('cas mismatch');
+            return cb(kvstore.ERROR_CAS_MISMATCH);
           }
           cb(err);
         }
