+++ lib/kvstore/memory.js
@@ -8,6 +8,7 @@
  */
 
 const Hapi = require('hapi');
+const kvstore = require('../kvstore');
 
 
 // Hapi's clone function only works on objects.
@@ -63,13 +64,13 @@ module.exports = {
       value = clone(value);
       process.nextTick(function() {
         if (data[key] === undefined) {
-          if (casid !== null) return cb('cas mismatch');
+          if (casid !== null) return cb(kvstore.ERROR_CAS_MISMATCH);
           data[key] = {
             value: value,
             casid: 1
           };
         } else {
-          if (data[key].casid !== casid)  return cb('cas mismatch');
+          if (data[key].casid !== casid)  return cb(kvstore.ERROR_CAS_MISMATCH);
           data[key].value = value;
           data[key].casid++;
         }
