+++ lib/kvstore/mysql.js
@@ -3,13 +3,14 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 /**
- * KVStore implementation backed by couchbase.
+ * KVStore implementation backed by mysql.
  *
  */
 
-const mysql = require('./mysql_wrapper.js');
+const mysql = require('../mysql/wrapper.js');
 const config = require('../config');
 const Hapi = require('hapi');
+const kvstore = require('../kvstore');
 
 var client;
 
@@ -28,7 +29,6 @@ module.exports = {
 
     cfg = Hapi.utils.merge(cfg, config.get('mysql'));
 
-    // mysql config requires
     var options = {};
 
     // Get the relevant client options from the mysql config
@@ -107,7 +107,7 @@ module.exports = {
       }
       client.query(query, args, function(err, result) {
         // XXX TODO: check for a constraint violation if casid == null.
-        if (casid !== null && result.affectedRows === 0) err = 'IncorrectCAS';
+        if (casid !== null && result.affectedRows === 0) err = kvstore.ERROR_CAS_MISMATCH;
         if (err) console.log(err);
         return cb(err);
       });
